chore(package): trim test fixture surface
