+++ package.json
@@ -10,7 +10,6 @@
   "files": [
     "dist",
     "src",
-    "tests/fixtures",
     "Dockerfile",
     ".dockerignore",
     "README.md"
@@ -53,10 +52,6 @@
     "./cma-sdk": {
       "types": "./dist/types/cma-sdk.d.ts",
       "default": "./src/cma-sdk.ts"
-    },
-    "./testing": {
-      "types": "./dist/types/testing.d.ts",
-      "default": "./src/testing.ts"
     }
   },
   "scripts": {
@@ -73,7 +68,6 @@
   "dependencies": {
     "@anthropic-ai/claude-agent-sdk": "^0.3.158",
     "@anthropic-ai/sdk": "^0.100.1",
-    "@cfworker/json-schema": "^4.1.1",
     "@modelcontextprotocol/client": "^2.0.0-alpha.2",
     "@orpc/client": "^1.14.3",
     "fflate": "^0.8.3",
